Removed videoURL from some of the models
