+++ Film/Network/API/Mocks/MockMovieInfoAPI.swift
@@ -14,7 +14,7 @@ class MockMovieInfoAPI: MovieInfoAPI {
     func getMovieInfo(movieId: Int, result: @escaping Handler<Movie>) {
         DispatchQueue.main.asyncAfter(deadline: .now() + simulatedDelay) {
             
-            let data: Movie = .init(id: 2, title: "El Camino", duration: 340, videoURL: "hehe", description: "Dopest movie ever made! Breaking bad typa stuff. Coming out October 11th, 2019.", poster: MockData.moviePosters[6], stoppedAt: 50)
+            let data: Movie = .init(id: 2, title: "El Camino", duration: 340, description: "Dopest movie ever made! Breaking bad typa stuff. Coming out October 11th, 2019.", poster: MockData.moviePosters[6], stoppedAt: 50)
             
             result(.success(data))
         }
